Quick Form
I made a quick user friendly form using html and javascript to download
the necessary information into a .txt file locally. My previous job use
to lose a lot of burs which are important for milling a lot of the times
no one would know why we lost them or how it broke so with this form it
helped.